+++ examples/data-sources/meraki_organization_extensions_thousand_eyes_networks/data-source.tf
@@ -1,0 +1,3 @@
+data "meraki_organization_extensions_thousand_eyes_networks" "example" {
+  organization_id = "123456"
+}