+++ examples/resources/meraki_organization_extensions_thousand_eyes_network/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_organization_extensions_thousand_eyes_network.example "<organization_id>,<id>"